requirements.txt
Django==4.1
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
psycopg2==2.9.7
requests~=2.31.0